move Check Descriptions to commit_status_helper
